Small refactorings
battle.py: in on_mous_press, we are using consistently the entity returned
by the grid.get_entity.
grid.py: the entity dictionary contains now a dictionary with the grid
coord as key and the entity reference as value. This allows for a quick
lookup of what is in what grid square.